@@ -12,29 +12,28 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^5.2.9",
-    "@angular/cdk": "^5.2.5",
-    "@angular/common": "5.2.9",
-    "@angular/compiler": "5.2.9",
-    "@angular/core": "5.2.9",
-    "@angular/flex-layout": "^5.0.0-beta.14",
-    "@angular/forms": "5.2.9",
-    "@angular/http": "5.2.9",
-    "@angular/material": "^5.2.5",
-    "@angular/platform-browser": "5.2.9",
-    "@angular/platform-browser-dynamic": "5.2.9",
-    "@angular/router": "5.2.9",
-    "angular-in-memory-web-api": "^0.5.3",
+    "@angular/animations": "6.0.1",
+    "@angular/cdk": "^6.0.1",
+    "@angular/common": "6.0.1",
+    "@angular/compiler": "6.0.1",
+    "@angular/core": "6.0.1",
+    "@angular/forms": "6.0.1",
+    "@angular/http": "6.0.1",
+    "@angular/material": "^6.0.1",
+    "@angular/platform-browser": "6.0.1",
+    "@angular/platform-browser-dynamic": "6.0.1",
+    "@angular/router": "6.0.1",
     "core-js": "^2.4.1",
     "hammerjs": "^2.0.8",
-    "ngx-infinite-scroll": "^0.8.4",
-    "rxjs": "^5.5.8",
-    "zone.js": "^0.8.24"
+    "rxjs": "^6.1.0",
+    "rxjs-tslint": "^0.1.3",
+    "zone.js": "^0.8.26"
   },
   "devDependencies": {
-    "@angular/cli": "1.7.3",
-    "@angular/compiler-cli": "5.2.9",
-    "@angular/language-service": "5.2.9",
+    "@angular-devkit/build-angular": "~0.6.1",
+    "@angular/cli": "^6.0.1",
+    "@angular/compiler-cli": "6.0.1",
+    "@angular/language-service": "6.0.1",
     "@types/jasmine": "~2.8.3",
     "@types/jasminewd2": "~2.0.2",
     "@types/node": "~6.0.60",
@@ -48,7 +47,7 @@
     "karma-jasmine-html-reporter": "^0.2.2",
     "protractor": "~5.1.2",
     "ts-node": "~4.1.0",
-    "tslint": "~5.9.1",
-    "typescript": "2.6.2"
+    "tslint": "^5.9.1",
+    "typescript": "2.7.2"
   }
 }
