@@ -1,11 +1,9 @@
 import { Injectable } from '@angular/core';
-import { User, Credentials } from '../user';
-import { Observable } from 'rxjs/Observable';
+import { User, Credentials, NewUser } from '../user';
+import { Observable , of } from 'rxjs';
 import { environment } from '../../environments/environment';
-import { HttpClient, HttpHeaders, HttpResponse, HttpErrorResponse } from '@angular/common/http';
+import { HttpClient, HttpHeaders, HttpResponse, HttpErrorResponse, HttpParams } from '@angular/common/http';
 import { tap, catchError } from 'rxjs/operators';
-import { of } from 'rxjs/observable/of';
-import { ErrorObservable } from 'rxjs/observable/ErrorObservable';
 import { Router } from '@angular/router';
 
 const httpOptions = {
@@ -19,6 +17,7 @@ const fullHttpOptions = {
 
 @Injectable()
 export class AuthService {
+    public isAuthenticated: boolean;
 
     constructor(private http: HttpClient, private router: Router) { }
 
@@ -28,26 +27,29 @@ export class AuthService {
 
     public setToken(token: string) {
         localStorage.setItem('auth-token', token);
+        this.isAuthenticated = true;
     }
 
     public removeToken() {
         localStorage.removeItem('auth-token');
+        this.isAuthenticated = false;
     }
 
-    public getCurrUser(): Observable<HttpResponse<User>> {
+    public getCurrUser(): Observable<User> {
         // return localStorage.getItem('auth-token') != null;
         const url = `${environment.apiUrl}/users/me`;
-        return this.http.get<HttpResponse<User>>(url, {observe: 'response'}).pipe(
+        return this.http.get<User>(url).pipe(
             tap((resp) => {
                 console.log(resp);
+                // this.setToken(resp.headers.get('Authorization'));
             }),
-            catchError(this.handleError)
+            catchError(this.handleError())
         );
     }
 
     // Returns the entire HTTP Response. JSON response
     // body is accessible via resp.body
-    public login(email: string, password: string): Observable<HttpResponse<User>> {
+    public login(email: string, password: string): Observable<HttpResponse<any>> {
         this.removeToken();
         const creds: Credentials = {email: email, password: password};
         const url = `${environment.apiUrl}/sessions`;
@@ -55,8 +57,9 @@ export class AuthService {
         return this.http.post<HttpResponse<User>>(url, creds, {observe: 'response'}).pipe(
             tap((resp) => {
                 this.successfulSignIn(resp);
-            }),
-            catchError(this.handleError)
+            })
+            // ,
+            // catchError(this.handleError())
         );
     }
 
@@ -71,7 +74,7 @@ export class AuthService {
                 this.router.navigate(['/login']);
                 console.log('signed out!');
             }),
-            catchError(this.handleError)
+            catchError(this.handleError())
         );
     }
 
@@ -79,42 +82,45 @@ export class AuthService {
         email: string,
         password: string,
         passwordConf: string,
-        userName: string,
         firstName: string,
-        lastName: string): Observable<HttpResponse<User>> {
+        lastName: string): Observable<HttpResponse<any>> {
         const url = `${environment.apiUrl}/users`;
-        const usr: User = {
+        const usr: NewUser = {
             email: email,
             password: password,
             passwordConf: passwordConf,
-            userName: userName,
             firstName: firstName,
             lastName: lastName
         };
         return this.http.post<HttpResponse<User>>(url, usr, {observe: 'response'})
-        .pipe(
-            tap((resp) => {
-                this.successfulSignIn(resp);
-            }),
-            catchError(this.handleError)
-        );
-    }
+            .pipe(
+                tap((resp) => {
+                    this.successfulSignIn(resp);
+                })
+            );
+}
 
     private successfulSignIn(resp) {
         this.setToken(resp.headers.get('Authorization'));
-        this.router.navigate(['/library']);
+        this.isAuthenticated = true;
+        this.router.navigate(['/browse']);
     }
 
     private log(message: string) {
+        // TODO: better logging, don't log to console
         console.log(message);
     }
 
-    private handleError(error: HttpErrorResponse) {
-        if (error.error instanceof ErrorEvent) {
-          // A client-side or network error occurred. Handle it accordingly.
-          console.error('An error occurred:', error.error.message);
-        }
-        // return an ErrorObservable with a user-facing error message
-        return new ErrorObservable(error);
+    private handleError<T> (operation = 'operation', result?: T) {
+        return (error: any): Observable<T> => {
+          // TODO: send the error to remote logging infrastructure
+          console.log('An error occurred:', error.error.message);
+
+          // TODO: better job of transforming error for user consumption
+          this.log(`${operation} failed: ${error.message}`);
+          this.removeToken();
+          // let the app keep running by returning an empty result
+          return of(result as T);
+        };
     }
 }
