@@ -1,8 +1,7 @@
 import { Injectable } from '@angular/core';
 import { HttpClient, HttpHeaders, HttpResponse } from '@angular/common/http';
 
-import { Observable } from 'rxjs/Observable';
-import { of } from 'rxjs/observable/of';
+import { Observable ,  of } from 'rxjs';
 import { catchError, map, tap } from 'rxjs/operators';
 
 
@@ -22,20 +21,21 @@ const textHttpOptions = {
 @Injectable()
 export class ShelfService {
 
-  private shelvesUrl = 'api/shelves'; // URL to web api
+  // private shelvesUrl = 'api/shelves'; // URL to web api
 
   constructor(private http: HttpClient) { }
 
   // GET shelves from the server
   getShelves (): Observable<Shelf[]> {
-    return this.http.get<Shelf[]>(this.shelvesUrl)
+    const url = `${environment.apiUrl}/shelves`;
+    return this.http.get<Shelf[]>(url)
       .pipe(
         tap(shelves => this.log(`fetched shelves`)),
         catchError(this.handleError('getShelves', []))
       );
   }
 
-  // GET the current users's shelves. Returns entire HTTP response
+  // GET the current users's shelves.
   getMyShelves(): Observable<Shelf[]> {
     const url = `${environment.apiUrl}/shelves/mine`;
     return this.http.get<Shelf[]>(url)
@@ -46,12 +46,31 @@ export class ShelfService {
       );
   }
 
+  getFeaturedShelves(): Observable<Shelf[]> {
+    const url = `${environment.apiUrl}/shelves/featured`;
+    return this.http.get<Shelf[]>(url)
+      .pipe(
+        tap((shelves) => {
+          console.log('fetched featured shelves');
+        })
+      );
+  }
+
   /** GET shelf by id. Will 404 if id not found */
-  getShelf(id: number): Observable<Shelf> {
-    const url = `${this.shelvesUrl}/${id}`;
+  getShelf(id: string): Observable<Shelf> {
+    const url = `${environment.apiUrl}/shelves/${id}`;
     return this.http.get<Shelf>(url).pipe(
       tap(_ => this.log(`fetched shelf id=${id}`)),
       catchError(this.handleError<Shelf>(`getShelf id=${id}`))
+    );
+  }
+
+  // GET a specified user's shelves
+  getUserShelves(userId: string): Observable<Shelf[]> {
+    const url = `${environment.apiUrl}/shelves/users/${userId}`;
+    return this.http.get<Shelf[]>(url).pipe(
+      tap(shelves => this.log(`fetched user's shelves`)),
+      catchError(this.handleError('getUserShelves', []))
     );
   }
 
@@ -61,7 +80,7 @@ export class ShelfService {
     const url = `${environment.apiUrl}/shelves/${shelf.id}`;
     return this.http.put(url, shelf, {headers: new HttpHeaders({ 'Content-Type': 'application/json' }), responseType: 'text'}).pipe(
       tap(_ => this.log(`updated shelf id=${shelf.id}`))
-      // ,catchError(this.handleError<any>('updateShelf'))
+      // ,catchError(this.handleError<any>('updateShelf')) // commented out so the subscriber can receive and process the error
     );
   }
 
@@ -70,16 +89,16 @@ export class ShelfService {
   addShelf(newShelf: NewShelf): Observable<Shelf> {
     const url = `${environment.apiUrl}/shelves`;
     return this.http.post<Shelf>(url, newShelf, httpOptions).pipe(
-      tap((returnedShelf: Shelf) => this.log(`added shelf w/ id=${returnedShelf.id}`)),
-      catchError(this.handleError<Shelf>('addShelf'))
+      tap((returnedShelf: Shelf) => this.log(`added shelf w/ id=${returnedShelf.id}`))
+      // , catchError(this.handleError<Shelf>('addShelf')) // commented out so the subscriber can receive and process the error
     );
   }
 
   // Delete: delete a shelf
   // api/shelves/{id}
-  deleteShelf(shelf: Shelf | number): Observable<Shelf> {
-    const id = typeof shelf === 'number' ? shelf : shelf.id;
-    const url = `${this.shelvesUrl}/$id`;
+  deleteShelf(shelf: Shelf | string): Observable<Shelf> {
+    const id = typeof shelf === 'string' ? shelf : shelf.id;
+    const url = `${environment.apiUrl}/shelves/${id}`;
 
     return this.http.delete<Shelf>(url, httpOptions).pipe(
       tap(_ => this.log(`deleted shelf id=${id}`)),
