@@ -1,27 +1,41 @@
 import { Routes } from '@angular/router';
-import { LibraryComponent } from './library/library.component';
 import { ReleaseDetailComponent } from './release/release-detail/release-detail.component';
 import { SigninComponent } from './auth/signin/signin.component';
 import { SignupComponent } from './auth/signup/signup.component';
-import { BrowseComponent } from './browse/browse.component';
+import { BrowseComponent, BrowseSubgroupComponent } from './browse/browse.module';
 import { BrowseSidenavComponent } from './browse-sidenav/browse-sidenav.module';
+import { ShelvesListComponent } from './shelves/shelves-list/shelves-list.component';
+import { SearchResultsComponent } from './search-results/search-results.component';
 
 export const SHELVES_ROUTES: Routes = [
-    {path: '', redirectTo: '/library', pathMatch: 'full'}, // default route redirect
-    {path: 'library', component: LibraryComponent}, // home page
+    {path: '', redirectTo: '/browse', pathMatch: 'full'}, // default route redirect
+    {path: 'library', redirectTo: '/browse', pathMatch: 'full'}, // home page
     {
         path: 'browse',
         component: BrowseSidenavComponent,
         children: [
-            {path: '', redirectTo: 'artists', pathMatch: 'full'},
+            {path: '', redirectTo: 'artists', pathMatch: 'full'}, // default to artists
+            {path: 'releases/:releaseId', component: ReleaseDetailComponent}, // release detail route
             {
-                path: ':id',
-                component: BrowseComponent
+                path: ':categoryId', // artists, genres, decades, etc
+                children: [
+                    {
+                        path: '', redirectTo: 'a', pathMatch: 'full'
+                    },
+                    {
+                        path: ':groupId',
+                        component: BrowseComponent,
+                    },
+                    {
+                        path: ':groupId/:subGroupId',
+                        component: BrowseSubgroupComponent // a browse subgroup
+                    }
+                ]
             }
         ],
     },
-    {path: 'shelves', component: LibraryComponent}, // shelves feature module
-    {path: 'library/releases/:releaseId', component: ReleaseDetailComponent}, // release detail route
+    {path: 'shelves', component: ShelvesListComponent}, // shelves feature module
+    {path: 'library/search/:query', component: SearchResultsComponent},
     {path: 'login', component: SigninComponent}, // sign in page
     {path: 'signup', component: SignupComponent},
     {path: '**', redirectTo: ''}
