@@ -1,8 +1,7 @@
-export class User {
+export class NewUser {
     email: string;
     password: string;
     passwordConf: string;
-    userName: string;
     firstName: string;
     lastName: string;
 }
@@ -12,4 +11,9 @@ export class Credentials {
     password: string;
 }
 
-export class NewUser {}
+export class User {
+    id: string;
+    email: string;
+    firstName: string;
+    lastName: string;
+}
